Changed text, fixed print stats not printing all stats, added add enemy feature.

+++ src/headers/Enemy.hpp
@@ -1,4 +1,3 @@
-#include<iostream>
 #include<string>
 
 using std::string;
@@ -7,18 +6,18 @@ class Enemy
 {
 public:
     Enemy(string name, short hp, short ac, short reflex, short fort, short will, short dc,
-            int damageBonus, int hitBonus) :
+            short damageBonus, short hitBonus) :
         _name(name), _maxHp(hp), _ac(ac), _reflex(reflex), _fort(fort), _will(will), _dc(dc),
         _damageBonus(damageBonus), _hitBonus(hitBonus)
         { _hp = _maxHp; };
     Enemy(){};
     void TakeDamage(short damage);
     void Heal(short damage);
-    void PrintStats() const; 
-    short GetHp() const { return _hp; } 
+    void PrintStats() const;
+    short GetHp() const { return _hp; }
     string GetName() const { return _name; };
-    int GetDamageBonus() const { return _damageBonus; }
-    int GetHitBonus() const { return _hitBonus; }
+    short GetDamageBonus() const { return _damageBonus; }
+    short GetHitBonus() const { return _hitBonus; }
 private:
     string _name;
     short _hp;
@@ -28,6 +27,6 @@ private:
     short _fort;
     short _will;
     short _dc;
-    int _damageBonus;
-    int _hitBonus;
+    short _damageBonus;
+    short _hitBonus;
 };
